requirements.txt
# requirements.txt
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
Pillow==10.0.0
numpy==1.24.3
databricks-sql-connector==2.9.3
pandas==2.0.3
opencv-python==4.8.0.74
scikit-image==0.21.0